Turns out get_beats is being called every time restart_interaction is called.
This was causing the rhythm file to be loaded over and over again.  It was
probably cached in memory by the OS anyway, but pre-load it once just to avoid
any unexpected overhead.

There's still some weird race condition where it can load the file a few times
at the beginning, but it's better than before at least.

Also, add some better cleanup at the end, in case a second rhythm game is
started after the first one finishes.